Baget_2024 version 2.7.20 (Customer & invoices)

+++ src/main/java/com/example/baget/buh/TransactionCollectiveInvoiceDTO.java
@@ -2,13 +2,14 @@ package com.example.baget.buh;
 
 import com.example.baget.orders.OrderPaySummaryDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record TransactionCollectiveInvoiceDTO(
         Long invoiceNo,
         List<OrderPaySummaryDTO> orders, // список замовлень у цьому рахунку
-        Double totalBilled,           // нараховано
-        Double totalPaid,             // сплачено
-        Double totalDue,              // залишок
-        Double totalCustomerBalance        // баланс клієнта
+        BigDecimal totalBilled,           // нараховано
+        BigDecimal totalPaid,             // сплачено
+        BigDecimal totalDue,              // залишок
+        BigDecimal totalCustomerBalance        // баланс клієнта
 ) {}
