+++ src/main/java/com/example/baget/buh/TransactionHistoryView.java
@@ -1,12 +1,13 @@
 package com.example.baget.buh;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public interface TransactionHistoryView {
     Long getTransactionId();
     OffsetDateTime getTransactionDate();
     TransactionType getTransactionType();
-    Double getAmount();
+    BigDecimal getAmount();
     String getReference();
     String getNote();
 }
