+++ src/main/java/com/example/baget/orders/OrderPaySummaryDTO.java
@@ -1,8 +1,10 @@
 package com.example.baget.orders;
 
+import java.math.BigDecimal;
+
 public record OrderPaySummaryDTO(
         Long orderNo,
-        Double billed,
-        Double paid,
-        Double due
+        BigDecimal billed,
+        BigDecimal paid,
+        BigDecimal due
 ) {}
