+++ src/main/java/com/example/baget/orders/OrderPrivateSummaryDTO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 @Getter
 @Setter
@@ -27,7 +28,7 @@ public class OrderPrivateSummaryDTO {
 
     private String empNo;
 
-    private Double itemsTotal;
+    private BigDecimal itemsTotal;
 
     private Integer statusOrder;
