+++ src/main/java/com/example/baget/orders/OrderPublicSummaryView.java
@@ -1,9 +1,10 @@
 package com.example.baget.orders;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public interface OrderPublicSummaryView {
     Long getOrderNo();     // Номер рахунку
     OffsetDateTime getSaleDate();     // Дата створення/виписки
-    Double getItemsTotal();   // Загальна сума рахунку
+    BigDecimal getItemsTotal();   // Загальна сума рахунку
 }
